requirements.txt
Flask==1.1.1
gevent==1.4.0
psycopg2==2.8.4
python_decouple==3.4
Flask-JWT-Extended==4.3.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0